+++ CHANGES.md
@@ -9,6 +9,8 @@ Unreleased
     context. This seems to fix resource warnings from sqlite3 during tests.
     {issue}`47`
 -   Add `path_separator` to config to suppress Alembic warning. {issue}`48`
+-   `upgrade` and `downgrade` check if the target is a revision before treating
+    it as a relative count. {issue}`45`
 
 ## Version 3.1.1